issue: 3782717 Move fifo into seperate source and header files
